Initial Proof-of-Concept
Still heavily in refinement. Definitely should be taken with a grain of salt at this point.

Long and short: Run the script, answer the prompts, and in response to "Data > " you may type "random" or give the name of a text file with country data in it. See (or use) monarchies.txt for an example of the required format.