BMP-format maps
After extensive trial-and-error testing, a rudimentary BMP-format mapping routine has been implemented. This is quite different from the previous attempt which output 6 different views of the planet; this routine instead unwraps the planet into a 2D array, distorting it heavily at the poles but showing the entire surface in one map.