A couple of quick changes.
Tapping enter on the main menu with no selection will now default to a new simulation.
compLangs() now divides its view into different screens which can be navigated in the same way as the matching function in gedReview() -- by extension, this means it will now display all languages instead of just the first handful.
Language.diff() is now a little more creative when performing a REPLACE operation. To be exact it uses a similar method to CCSCommon.namecheck() with its letter grouping.